radar testing WIP
- still not working
-added stuff for debugbing
